+++ src/main/java/ladder/domain/ladder/Line.java
@@ -1,9 +1,8 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.IntStream;
 
 import static java.util.stream.IntStream.*;
