+++ src/main/java/ladder/domain/ladder/Position.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 import java.util.Objects;
