+++ src/main/java/ladder/domain/user/Players.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.user;
 
 import java.util.List;
 import java.util.Objects;
