+++ src/main/java/ladder/view/ResultView.java
@@ -1,7 +1,10 @@
 package ladder.view;
 
-import ladder.domain.*;
-import ladder.domain.Position;
+import ladder.domain.ladder.Ladder;
+import ladder.domain.ladder.Line;
+import ladder.domain.ladder.Point;
+import ladder.domain.user.Player;
+import ladder.domain.user.Players;
 
 import java.util.List;
 
@@ -11,18 +14,24 @@ public class ResultView {
         throw new AssertionError();
     }
 
-    public static void printLadder(Players players, Height height) {
-        Ladder ladder = Ladder.of(players, height);
+    public static void printPlayerName(Players players) {
         System.out.println("실행결과");
         System.out.println();
+        List<Player> users = players.getPlayers();
+        System.out.printf("%-5s", users.get(0));
+        for (int i = 1; i < users.size() - 1; i++) {
+            System.out.printf("%6s", users.get(i));
+        }
+        System.out.printf("%6s%n", users.get(users.size()-1));
+    }
 
-        printPlayerName(players);
+    public static void printLadder(Ladder ladder) {
 
         List<Line> lines = ladder.getLines();
         for (Line line : lines) {
             List<Point> points = line.getPoints();
             System.out.print("     |");
-            for (int i = 1; i < points.size() -1; i++) {
+            for (int i = 1; i < points.size(); i++) {
                 if (points.get(i).isRight()) {
                     System.out.print("-----|");
                 }
@@ -32,15 +41,6 @@ public class ResultView {
             }
             System.out.println();
         }
-    }
-
-    private static void printPlayerName(Players players) {
-        List<Player> users = players.getPlayers();
-        System.out.printf("%-5s", users.get(0));
-        for (int i = 1; i < users.size() - 1; i++) {
-            System.out.printf("%6s", users.get(i));
-        }
-        System.out.printf("%6s%n", users.get(users.size()-1));
     }
 
 }
