+++ src/test/java/ladder/domain/user/PlayerNameTest.java
@@ -1,6 +1,6 @@
-package ladder.domain;
+package ladder.domain.user;
 
-import ladder.domain.PlayerName;
+import ladder.domain.user.PlayerName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
