+++ src/test/java/ladder/domain/user/PlayersTest.java
@@ -1,5 +1,8 @@
-package ladder.domain;
+package ladder.domain.user;
 
+import ladder.domain.user.Player;
+import ladder.domain.user.PlayerName;
+import ladder.domain.user.Players;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
