+++ src/test/java/ladder/view/ResultViewTest.java
@@ -1,8 +1,10 @@
 package ladder.view;
 
-import ladder.domain.*;
-import ladder.domain.PlayerName;
-import ladder.domain.Position;
+import ladder.domain.ladder.Height;
+import ladder.domain.ladder.Ladder;
+import ladder.domain.user.Player;
+import ladder.domain.user.PlayerName;
+import ladder.domain.user.Players;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
@@ -19,9 +21,9 @@ class ResultViewTest {
         Players players = new Players(Arrays.asList(player1, player2, player3, player4, player5));
 
         Height height = new Height(4);
-        Ladder ladder = LadderGenerator.createLadder(players, height);
-
-        ResultView.printLadder(players, height);
+        Ladder ladder = Ladder.createLadder(players, height);
+        ResultView.printPlayerName(players);
+        ResultView.printLadder(ladder);
     }
 
 }
